pyproject.toml
[project]
name = "clashbot-nav"
version = "0.1.0"
description = ""
authors = [
    {name = "matt-lenovo-laptop",email = "matmigg0804@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "numpy (>=2.2,<3.0)",
    "pygetwindow (>=0.0.9,<0.0.10)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
